pyproject.toml
[tool.poetry]
name = "mqtt-hello"
version = "0.1.0"
description = ""
authors = ["ricklon <rianders@docs.rutgers.edu>"]
readme = "README.md"
packages = [{include = "mqtt_hello"}]

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
